[brief] Update the install/clean scripts.
[detailed]
- Ensure that zshenv doesn't modify the XDG folders if they're already
  set.

+++ clean.sh
@@ -1,18 +1,29 @@
 #!/bin/sh
 
-rm "$HOME/.zshrc"
-rm "$HOME/.zprofile"
+cur_os=$(uname)
+
+# ZSH
 rm "$HOME/.zshenv"
-rm "$HOME/.zsh_plugins.txt"
-rm "$HOME/.zsh_plugins.zsh"
-rm "$HOME/.p10k.zsh"
-rm "$HOME/.zshfn"
-rm "$HOME/.zshpy"
+rm "$HOME/.config/zsh"
+
+# Git
 rm "$HOME/.gitconfig"
 rm "$HOME/.gitignore"
 rm "$HOME/.gittemplate.txt"
+
+# Nvim
 rm "$HOME/.config/nvim"
+
+# Language files
 rm "$HOME/.clang-format"
-rm "$HOME/.tmux.conf"
+
+# Bat
 rm "$HOME/.config/bat"
+
+# WezTerm
 rm "$HOME/.config/wezterm"
+
+if [[ "$cur_os" != "Darwin" ]]; then
+    rm "$HOME/.tmux.conf"
+    rm -rf "$HOME/.tmux"
+fi
